Move RealmContextType in another file
